+++ src/api/mod.rs
@@ -9,6 +9,7 @@
 
 pub mod bookmark_ops;
 pub mod channel_ops;
+pub mod chat_ops;
 mod client;
 pub mod edge;
 pub mod identity_ops;